Distinguish the location of the test archive with "test_storage".
Before this change 2 storage options were supported: if the option
"test_path" was set, a ready-to-run test was expected in the specified
folder. Otherwise the options "archive_path" and "test_archive"
specified a compressed test, which was extracted to the folder
"depack_path".
Now there is a new option "test_storage", which can have one of the
following values: "FOLDER" and "LOCAL_ARCHIVE". "FOLDER" is the
ready-to-run test and "LOCAL_ARCHIVE" the compressed test.